+++ scripts/freebsd13/pkg.sh
@@ -7,7 +7,7 @@ export ASSUME_ALWAYS_YES=yes
 
 # We need to use HTTP until the CA bundle has been updated.
 mkdir -p /usr/local/etc/pkg/repos/
-rm /var/db/pkg/FreeBSD.meta /var/db/pkg/repo-FreeBSD.sqlite
+rm -f /var/db/pkg/FreeBSD.meta /var/db/pkg/repo-FreeBSD.sqlite /var/db/pkg/repo-FreeBSD.sqlite-journal
 echo 'FreeBSD: { url: "pkg+http://pkg.FreeBSD.org/${ABI}/latest" }' > /usr/local/etc/pkg/repos/FreeBSD.conf
 
 # Install the packages needed to update the CA bundle.
@@ -16,24 +16,26 @@ pkg-static update -f
 pkg-static upgrade --yes perl5 p5-MIME-Base64 p5-Carp curl ca_root_nss
 
 # Download the bundle generator.
-curl --silent --location --output $HOME/mk-ca-bundle.pl https://raw.githubusercontent.com/curl/curl/85f91248cffb22d151d5983c32f0dbf6b1de572a/lib/mk-ca-bundle.pl
-sha256 -c f819e5844935bad3d7eebab566b55066f21bd7138097d2baab7842bd04fd92d2 $HOME/mk-ca-bundle.pl || exit 1
-chmod +x $HOME/mk-ca-bundle.pl
+# curl --silent --location --output $HOME/mk-ca-bundle.pl https://raw.githubusercontent.com/curl/curl/85f91248cffb22d151d5983c32f0dbf6b1de572a/lib/mk-ca-bundle.pl
+# sha256 -c f819e5844935bad3d7eebab566b55066f21bd7138097d2baab7842bd04fd92d2 $HOME/mk-ca-bundle.pl || exit 1
+# chmod +x $HOME/mk-ca-bundle.pl
 
 # To ensure we save the certdata.txt file in a predictable place we change directory first.
-cd $HOME && $HOME/mk-ca-bundle.pl $HOME/ca-bundle.crt
+# cd $HOME && $HOME/mk-ca-bundle.pl $HOME/ca-bundle.crt
 
 # Move the updated bundle to all the places it might be needed.
-cp $HOME/ca-bundle.crt /etc/ssl/cert.pem
-cp $HOME/ca-bundle.crt /usr/local/openssl/cert.pem
-cp $HOME/ca-bundle.crt /usr/local/etc/ssl/cert.pem
-cp $HOME/ca-bundle.crt /usr/local/share/certs/ca-root-nss.crt
+# cp $HOME/ca-bundle.crt /etc/ssl/cert.pem
+# cp $HOME/ca-bundle.crt /usr/local/openssl/cert.pem
+# cp $HOME/ca-bundle.crt /usr/local/etc/ssl/cert.pem
+# cp $HOME/ca-bundle.crt /usr/local/share/certs/ca-root-nss.crt
 
 # Cleanup the downloaded files and clear the cached repo data.
-rm $HOME/ca-bundle.crt $HOME/certdata.txt $HOME/mk-ca-bundle.pl /var/db/pkg/FreeBSD.meta /var/db/pkg/repo-FreeBSD.sqlite
+# rm $HOME/ca-bundle.crt $HOME/certdata.txt $HOME/mk-ca-bundle.pl /var/db/pkg/FreeBSD.meta /var/db/pkg/repo-FreeBSD.sqlite
 
 # Switch to using HTTPS and perform the system upgrade.
+rm -f /var/db/pkg/FreeBSD.meta /var/db/pkg/repo-FreeBSD.sqlite /var/db/pkg/repo-FreeBSD.sqlite-journal
 echo 'FreeBSD: { url: "pkg+https://pkg.FreeBSD.org/${ABI}/latest" }' > /usr/local/etc/pkg/repos/FreeBSD.conf
+
 pkg-static update -f
 pkg-static upgrade --yes
 
@@ -41,7 +43,8 @@ pkg-static upgrade --yes
 pkg-static upgrade --yes vim curl wget sudo bash gnuls gnugrep psmisc
 
 # Since most scripts expect bash to be in the bin directory, create a symlink.
-ln -s /usr/local/bin/bash /bin/bash
+[ ! -f /bin/bash ] && [ -f  /usr/local/bin/bash ] && ln -s /usr/local/bin/bash /bin/bash
+[ ! -f /usr/bin/bash ] && [ -f  /usr/local/bin/bash ] && ln -s /usr/local/bin/bash /usr/bin/bash
 
 # Disable fortunate cookies.
 sed -i "" -e "/fortune/d" /usr/share/skel/dot.login
