+++ scripts/openbsd7/pkg.sh
@@ -12,7 +12,8 @@ pkg_add -u
 pkg_add -I curl wget bash sudo-- vim--no_x11
 
 # Since most scripts expect bash to be in the bin directory, create a symlink.
-ln -s /usr/local/bin/bash /bin/bash
+[ ! -f /bin/bash ] && [ -f  /usr/local/bin/bash ] && ln -s /usr/local/bin/bash /bin/bash
+[ ! -f /usr/bin/bash ] && [ -f  /usr/local/bin/bash ] && ln -s /usr/local/bin/bash /usr/bin/bash
 
 # Some hypervisors require this to run OpenBSD properly.
 echo "kern.allowkmem=1" > /etc/sysctl.conf
